修复全局配置非单例bug
