+++ jrasp-core/src/main/java/com/jrasp/agent/core/CoreConfigure.java
@@ -1,6 +1,5 @@
 package com.jrasp.agent.core;
 
-import com.jrasp.agent.api.RaspConfig;
 import com.jrasp.agent.api.annotation.Information;
 import com.jrasp.agent.core.util.FeatureCodec;
 import com.jrasp.agent.core.util.ProcessHelper;
@@ -9,6 +8,7 @@ import com.jrasp.agent.core.util.string.RaspStringUtils;
 import org.apache.commons.io.FileUtils;
 
 import java.io.File;
+import java.lang.instrument.Instrumentation;
 import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
@@ -18,7 +18,7 @@ import java.util.Map;
  * 内核启动配置
  * Created by luanjia@taobao.com on 16/10/2.
  */
-public class CoreConfigure implements RaspConfig {
+public class CoreConfigure  {
 
     private static final String KEY_NAMESPACE = "namespace";
     private static final String DEFAULT_VAL_NAMESPACE = "default";
@@ -41,19 +41,6 @@ public class CoreConfigure implements RaspConfig {
     // 技术支持链接
     public static final String JRASP_SUPPORT_URL = "https://www.jrasp.com";
 
-    // go viper 不区分大小写, java为了便于区分，这里使用下划线命名规则
-    private volatile int block_status_code = 302;
-
-    private volatile boolean check_disable = false;
-
-    private volatile String redirect_url = "https://www.jrasp.com/block.html";
-
-    private volatile String json_block_content = "{\"error\":true, \"reason\": \"Request blocked by JRASP (https://www.jrasp.com)\"}";
-
-    private volatile String xml_block_content = "<?xml version=\"1.0\"?><doc><error>true</error><reason>Request blocked by JRASP</reason></doc>";
-
-    private volatile String html_block_content = "</script><script>location.href=\"https://www.jrasp.com/block.html\"</script>";
-
     private static final FeatureCodec codec = new FeatureCodec(';', '=');
 
     private final Map<String, String> featureMap = new LinkedHashMap<String, String>();
@@ -67,14 +54,14 @@ public class CoreConfigure implements RaspConfig {
         return codec.toMap(featureString);
     }
 
-    private static volatile CoreConfigure instance;
-
+    /**
+     * @see com.jrasp.agent.launcher110.AgentLauncher#install(Map, Instrumentation) 被反射初始化
+     * 即每次执行 attach 生成一个新的对象
+     * @param featureString
+     * @return
+     */
     public static CoreConfigure toConfigure(final String featureString) {
-        return instance = new CoreConfigure(featureString);
-    }
-
-    public static CoreConfigure getInstance() {
-        return instance;
+        return new CoreConfigure(featureString);
     }
 
     /**
@@ -207,66 +194,6 @@ public class CoreConfigure implements RaspConfig {
     //  获取进程运行时pid/
     public String getRuntimeTokenPath() {
         return getProcessPidPath() + File.separatorChar + TOKEN_FILE_NAME;
-    }
-
-    @Override
-    public boolean isCheckDisable() {
-        return check_disable;
-    }
-
-    @Override
-    public void setCheckDisable(boolean checkDisable) {
-        this.check_disable = checkDisable;
-    }
-
-    @Override
-    public String getRedirectUrl() {
-        return this.redirect_url;
-    }
-
-    @Override
-    public String getJsonBlockContent() {
-        return json_block_content;
-    }
-
-    @Override
-    public String getXmlBlockContent() {
-        return xml_block_content;
-    }
-
-    @Override
-    public String getHtmlBlockContent() {
-        return html_block_content;
-    }
-
-    @Override
-    public void setRedirectUrl(String redirectUrl) {
-        this.redirect_url = redirectUrl;
-    }
-
-    @Override
-    public void setJsonBlockContent(String jsonBlockContent) {
-        this.json_block_content = jsonBlockContent;
-    }
-
-    @Override
-    public void setXmlBlockContent(String xmlBlockContent) {
-        this.xml_block_content = xmlBlockContent;
-    }
-
-    @Override
-    public void setHtmlBlockContent(String htmlBlockContent) {
-        this.html_block_content = htmlBlockContent;
-    }
-
-    @Override
-    public int getBlockStatusCode() {
-        return block_status_code;
-    }
-
-    @Override
-    public void setBlockStatusCode(int blockStatusCode) {
-        this.block_status_code = blockStatusCode;
     }
 
 }
